scanner

+++ fahrenheit/Fahrenheit.java
@@ -1,7 +1,7 @@
 public class Fahrenheit {
     public static void main(String[] args) {
-        int Fahrenheit = 80;
-       double Celsius = (5.0/9.0)*(Fahrenheit - 32);
+        float Fahrenheit = 80;
+       float Celsius = (5.0f/9.0f)*(Fahrenheit - 32);
        
         System.out.println(Celsius);
     }
